requirements.txt
# Core ML libraries
torch>=1.9.0
transformers>=4.20.0
sentence-transformers>=2.2.0
datasets>=2.0.0

# Vector search and indexing
faiss-cpu>=1.7.0
chromadb>=0.3.0

# Data processing
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Web framework
streamlit>=1.25.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Language models and APIs
openai>=0.27.0
huggingface-hub>=0.14.0

# Evaluation frameworks
ragas>=0.0.16
datasets>=2.0.0

# Text processing
textstat>=0.7.0
spacy>=3.4.0

# Utilities
tqdm>=4.62.0
python-dotenv>=0.19.0
jsonlines>=3.0.0

# Development tools
jupyter>=1.0.0
notebook>=6.4.0

# Optional: GPU support (uncomment if using CUDA)
# torch>=1.9.0+cu111 -f https://download.pytorch.org/whl/torch_stable.html
# faiss-gpu>=1.7.0